LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/navigate-history-back-noop-expected.txt



FAIL history.back() does not fire a navigate event when there's nothing to navigate assert_equals: expected 2 but got 0
